+++ src/age_qt_gui/age_ui_qt_settings.hpp
@@ -452,7 +452,7 @@ signals:
 public slots:
 
     void audio_output_activated(QAudioDeviceInfo device, QAudioFormat format, int buffer_size, int downsampler_fir_size);
-    void set_emulator_screen_size(uint width, uint height);
+    void set_emulator_screen_size(int16_t width, int16_t height);
